WIP, most of a saved CR now rendering correctly

+++ src/migrations/20250905180510-create-collab-report-tables.js
@@ -43,7 +43,7 @@ module.exports = {
           allowNull: false,
         },
         duration: {
-          type: Sequelize.SMALLINT,
+          type: Sequelize.FLOAT,
           allowNull: false,
           validate: {
             min: 0,
